Adding a sort to guarantee stable parameters
Marinator had a very subtle bug previously. The order of the parameters for the MarinadeHelper and the constructor were determined by file traversal order. It's possible that this is different across different build environments, which leads to compilation errors.

The fix is to sort the parameters in a deterministic order. This allows us to guarantee the order of the constructor params.

Additionally, this PR adds the correct attribution for third-party libraries used in Marinator.