Updating Marinator infrastructure.
Marinator is getting a bit of an update. This upgrade does two major things:
1) Sets the library to use java 1.8 compatibility.
2) Upgrade Marinator to use the latest Android grade plugin/gradle versions.
3) Marks annotation processors as incremental. This will give us better performance in newer build chains.

Most of this is just boilerplate generated by updating the gradle wrapper. The only substantive pieces here adding the new META-INF file to declare the processor as incremental, and updating the gradle files themselves to use the new plugins.

Once this is merged, this will be packaged and released as v1.1.0.

[no-jira] Tech update